refactor: Corrections in extractor class.
