+++ src/features/extractor.py
@@ -1,4 +1,9 @@
-class Extractor():
+from datasets.arrow_dataset import Dataset
+
+import torch
+
+
+class Extractor(torch.utils.data.Dataset):
     """
     Extractor
     =======
@@ -9,8 +14,22 @@ class Extractor():
         name (str): The name of the extractor. Default: "EmptyExtractor".
     """
 
-    def __init__(self, name="EmptyExtractor"):
+    def __init__(self, dataset: Dataset, column: str, name="Extractor"):
         """
         Initializes a new Extractor
         """
         self.name = name
+        self.dataset = dataset
+        self.column = column
+
+        # Define the device
+        self.device = torch.device(
+            'cuda' if torch.cuda.is_available() else 'cpu')
+
+    def __len__(self):
+        return len(self.dataset)
+
+    def __getitem__(self, idx):
+        # Get the data
+        item = self.dataset[idx][self.column]
+        return item
