+++ src/features/waveform_extractor.py
@@ -4,7 +4,7 @@ from datasets.arrow_dataset import Dataset
 from .extractor import Extractor
 
 
-class WaveformExtractor(torch.utils.data.Dataset, Extractor):
+class WaveformExtractor(Extractor):
     """
     WaveformExtractor
     =======
@@ -18,25 +18,16 @@ class WaveformExtractor(torch.utils.data.Dataset, Extractor):
         name (str): The name of the extractor
     """
 
-    def __init__(self, dataset: Dataset, audio_column="audio", time_fixed=0, name="Waveform", ):
+    def __init__(self, dataset: Dataset, column="audio", time_fixed=0):
         """Initializes a new WaveformExtractor
         """
-        Extractor.__init__(self, name)
+        super().__init__(dataset=dataset, column=column, name="Waveform")
 
-        self.dataset = dataset
-        self.audio_column = audio_column
         self.time_fixed = time_fixed
-
-        # Define the device
-        self.device = torch.device(
-            'cuda' if torch.cuda.is_available() else 'cpu')
-
-    def __len__(self):
-        return len(self.dataset)
 
     def __getitem__(self, idx):
         # Get the audio data
-        audio_data = self.dataset[idx][self.audio_column]
+        audio_data = super().__getitem__(idx)
 
         # Audio Data Keys = 'path', 'array', and 'sampling_rate'
         waveform = audio_data['array']
